added tests for site user id getter; restructured tests to separate unit tests from integration tests
