+++ .mocharc.json
@@ -1,4 +1,5 @@
 {
     "$schema": "https://json.schemastore.org/mocharc.json",
-    "bail": true
+    "bail": true,
+    "recursive": true
 }
