+++ test/unit/package.js
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { parsePackage, parsePerson } from "../src/utils/package.js";
+import { parsePackage, parsePerson } from "../../src/utils/package.js";
 
 describe('Package Parsing', () => {
     describe('parsePerson', () => {
